requirements.txt
Flask==2.3.2
Flask-WTF==1.2.1
requests==2.31.0
libtorrent
humanize==4.9.0
werkzeug==2.3.7
speedtest-cli==2.1.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-Admin==1.6.1
python-dotenv==1.0.0
